refactor(telegram): move the unified delivery path to delivery.rs (#471 phase 4)
deliver_final_response and drain_remaining_display move to a dedicated
delivery.rs (642 lines): the ONE path every Telegram turn ends on, for
live and resumed turns alike. Pure move with import/visibility plumbing
only (fire_reaction widened to pub(crate)).

handler.rs lands at 3983 lines — from 7221 at the start of #471.
